order_service/requirements.txt
PyMySQL==1.1.0
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
gunicorn==21.2.0
cryptography>=3.4.8
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
